refactor: Move all server related code into one file.
