+++ app/root.tsx
@@ -13,8 +13,8 @@ import { type ReactNode, useEffect } from 'react';
 
 import { GeneralErrorBoundary } from '#components/error-boundary';
 import UI from '#components/ui';
+import { getUser } from '#utils/auth/auth.server.ts';
 import { useAuthBroadcast } from '#utils/auth/user';
-import { getUser } from '#utils/auth/utils.server.ts';
 import { combineHeaders } from '#utils/misc.ts';
 import { ClientHintsFallback } from '#utils/theme/client-hints-fallback';
 import { getHints } from '#utils/theme/client-hints.server';
