+++ app/routes/_foh+/_auth+/login.tsx
@@ -7,8 +7,7 @@ import { useActionData, useSubmit } from '@remix-run/react';
 import { Form } from 'react-aria-components';
 
 import UI from '#components/ui';
-import { signup } from '#utils/auth/auth.server.ts';
-import { requireAnonymous } from '#utils/auth/utils.server.ts';
+import { requireAnonymous, signup } from '#utils/auth/auth.server.ts';
 
 export const handle = {
   pageTitle: 'Log In',
