+++ app/routes/sse+/task.$taskId.tsx
@@ -1,6 +1,6 @@
 import type { LoaderFunctionArgs } from '@remix-run/node';
 import { eventStream } from 'remix-utils/sse/server';
-import { requireAdmin } from '#utils/auth/utils.server.js';
+import { requireAdmin } from '#utils/auth/auth.server.ts';
 import {
   type TaskProgressEvent,
   taskProgressEventBus,
